+++ packages/better-uptime/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.21.2
+
+### Patch Changes
+
+Update pulumi-better-uptime from 0.21.2 to 0.21.3
+
+**Full Changelog**: https://github.com/BetterStackHQ/terraform-provider-better-uptime/compare/v0.21.2...v0.21.3
+
 ## 0.21.1
 
 ### Patch Changes