+++ packages/bunnynet/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.14.4
+
+### Patch Changes
+
+Update pulumi-bunnynet from 0.14.2 to 0.14.3
+
+**Full Changelog**: https://github.com/BunnyWay/terraform-provider-bunnynet/compare/v0.14.2...v0.14.3
+
 ## 0.14.3
 
 ### Patch Changes